refactor structured model json output, html rendered table output
